# artifacts/api-server/src/routes/announcements.ts
import { Router, type IRouter } from "express";
import { eq, desc } from "drizzle-orm";
import { db, announcementsTable } from "@workspace/db";
import {
  ListAnnouncementsQueryParams,
  GetAnnouncementParams,
  ListAnnouncementsResponse,
  GetAnnouncementResponse,
} from "@workspace/api-zod";
import {
  makePageMeta,
  buildSearchCondition,
  combine,
  countSql,
} from "./lib/pagination";

const router: IRouter = Router();

router.get("/announcements", async (req, res): Promise<void> => {
  const parsed = ListAnnouncementsQueryParams.safeParse(req.query);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }
  const { q, year, category, page = 1, pageSize = 12 } = parsed.data;
  const where = combine(
    buildSearchCondition(q, [announcementsTable.title, announcementsTable.summary, announcementsTable.body]),
    year ? eq(announcementsTable.year, year) : undefined,
    category ? eq(announcementsTable.category, category) : undefined,
  );
  const items = await db
    .select()
    .from(announcementsTable)
    .where(where)
    .orderBy(desc(announcementsTable.publishedAt))
    .limit(pageSize)
    .offset((page - 1) * pageSize);
  const [{ total } = { total: 0 }] = await db
    .select({ total: countSql })
    .from(announcementsTable)
    .where(where);
  const yearsRows = await db
    .select({ year: announcementsTable.year, count: countSql })
    .from(announcementsTable)
    .groupBy(announcementsTable.year)
    .orderBy(desc(announcementsTable.year));
  const catsRows = await db
    .selectDistinct({ category: announcementsTable.category })
    .from(announcementsTable);
  res.json(
    ListAnnouncementsResponse.parse({
      items,
      pagination: makePageMeta(page, pageSize, Number(total)),
      availableYears: yearsRows.map((r) => ({ year: r.year, count: Number(r.count) })),
      availableCategories: catsRows.map((r) => r.category),
    }),
  );
});

router.get("/announcements/:id", async (req, res): Promise<void> => {
  const params = GetAnnouncementParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }
  const [row] = await db.select().from(announcementsTable).where(eq(announcementsTable.id, params.data.id));
  if (!row) {
    res.status(404).json({ error: "Not found" });
    return;
  }
  res.json(GetAnnouncementResponse.parse(row));
});

export default router;