# lib/api-zod/src/generated/types/announcementItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API for the municipality citizen portal
 * OpenAPI spec version: 0.1.0
 */

export interface AnnouncementItem {
  id: number;
  title: string;
  summary: string;
  body: string;
  category: string;
  publishedAt: Date;
  year: number;
}