# lib/api-zod/src/generated/types/gazetteItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API for the municipality citizen portal
 * OpenAPI spec version: 0.1.0
 */

export interface GazetteItem {
  id: number;
  issueNumber: string;
  title: string;
  summary: string;
  publishedAt: Date;
  year: number;
  fileUrl: string;
}