# lib/api-zod/src/generated/types/projectListPage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API for the municipality citizen portal
 * OpenAPI spec version: 0.1.0
 */
import type { Pagination } from "./pagination";
import type { ProjectItem } from "./projectItem";
import type { YearCount } from "./yearCount";

export interface ProjectListPage {
  items: ProjectItem[];
  pagination: Pagination;
  availableYears: YearCount[];
  availableCategories: string[];
  availableStatuses: string[];
}