# lib/api-zod/src/generated/types/tenderItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API for the municipality citizen portal
 * OpenAPI spec version: 0.1.0
 */
import type { TenderItemStatus } from "./tenderItemStatus";

export interface TenderItem {
  id: number;
  title: string;
  summary: string;
  body: string;
  status: TenderItemStatus;
  publishedAt: Date;
  deadline: Date;
  year: number;
}